requirements.txt
absl-py==1.0.0
backcall==0.2.0
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.6
debugpy==1.4.1
decorator==5.0.9
entrypoints==0.4
fonttools==4.33.3
future==0.18.2
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.13.0
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.3.1
jupyter-core==4.10.0
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
mish-cuda==0.0.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
nest-asyncio==1.5.5
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
seaborn==0.11.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.0.31.post2005241907
torch==1.7.0
torchaudio==0.7.0a0+ac17b64
torchvision==0.8.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
